Only rebuild xcbeautify if it doesn't exist at ./xcbeautify

+++ Makefile
@@ -58,10 +58,11 @@ test-fast:
 	swift test --disable-automatic-resolution
 
 xcbeautify:
-	rm -rf .build/checkouts/xcbeautify
-	git clone --depth=1 https://github.com/thii/xcbeautify.git .build/checkouts/xcbeautify
-	cd .build/checkouts/xcbeautify && make build
-	binpath=`cd .build/checkouts/xcbeautify && swift build -c release --show-bin-path` && ln -sf $$binpath/xcbeautify
+	if [[ ! -f ./xcbeautify ]]; then
+		git clone --depth=1 https://github.com/thii/xcbeautify.git .build/checkouts/xcbeautify
+		cd .build/checkouts/xcbeautify && make build
+		binpath=`cd .build/checkouts/xcbeautify && swift build -c release --show-bin-path` && ln -sf $$binpath/xcbeautify
+	fi
 
 docker-build: version
 	docker build -t $(DOCKER_IMAGE):$(VERSION) .
